core: fix clippy and `no_std` build

+++ emissary-core/src/i2np/delivery_status.rs
@@ -45,7 +45,7 @@ impl DeliveryStatus {
             rest,
             Self {
                 message_id,
-                timestamp: Duration::from_millis(timestamp as u64),
+                timestamp: Duration::from_millis(timestamp),
             },
         ))
     }
